modules/grpc-DW-proxy/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
kafka-python==2.1.3